[JBERET-286] Abstract the JobOperatorImpl so it can be extended by implementations. Added a JobOperatorContext and context selector to allow the environment to control the JobOperator used. Use a new DelegatingJobOperator as the default implemented returned by the BatchRuntime.getJobOperator().
