+++ jberet-core/src/main/resources/META-INF/services/javax.batch.operations.JobOperator
@@ -1,2 +1,2 @@
-org.jberet.operations.JobOperatorImpl
+org.jberet.operations.DelegatingJobOperator
